--- a/fastapi_d_reflowpro/.env.example
+++ b/fastapi_d_reflowpro/.env.example
@@ -1,22 +1,22 @@
 # Database Configuration
-DATABASE_URL=postgresql+asyncpg://postgres:postgres@localhost:5432/dreflowpro
+DATABASE_URL=postgresql+asyncpg://username:password@localhost:5432/dreflowpro
 
 # Redis Configuration
 REDIS_URL=redis://localhost:6379/0
 
 # Security
-SECRET_KEY=your_secret_key_here
+SECRET_KEY=REPLACE_WITH_YOUR_SECRET_KEY_HERE
 DEBUG=true
 
 # OAuth Configuration (Optional)
-GOOGLE_CLIENT_ID=your_google_client_id
-GOOGLE_CLIENT_SECRET=your_google_client_secret
+GOOGLE_CLIENT_ID=REPLACE_WITH_GOOGLE_CLIENT_ID
+GOOGLE_CLIENT_SECRET=REPLACE_WITH_GOOGLE_CLIENT_SECRET
 
-GITHUB_CLIENT_ID=your_github_client_id
-GITHUB_CLIENT_SECRET=your_github_client_secret
+GITHUB_CLIENT_ID=REPLACE_WITH_GITHUB_CLIENT_ID
+GITHUB_CLIENT_SECRET=REPLACE_WITH_GITHUB_CLIENT_SECRET
 
-MICROSOFT_CLIENT_ID=your_microsoft_client_id
-MICROSOFT_CLIENT_SECRET=your_microsoft_client_secret
+MICROSOFT_CLIENT_ID=REPLACE_WITH_MICROSOFT_CLIENT_ID
+MICROSOFT_CLIENT_SECRET=REPLACE_WITH_MICROSOFT_CLIENT_SECRET
 
 # CORS Origins
 CORS_ORIGINS=["http://localhost:3000", "http://localhost:3001"]